requirements.txt
streamlit
pandas
openai
plotly
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
